LEKSE-88 set state when sending msg
