+++ src/styles/Chat.less
@@ -1,9 +1,10 @@
 @import "colors";
 
 .chat {
-  display: block;
+  display: flex;
+  flex-direction: column;
   margin: 0 auto;
   background-color: white;
-  height: 100%;
+  height: 90%;
   width: 90%;
 }
